// src/types/task.ts
export type TaskCategory = 
  | 'presentation' 
  | 'assignment' 
  | 'quiz' 
  | 'lab-report' 
  | 'lab-final' 
  | 'task' 
  | 'documents'
  | 'blc'
  | 'groups'
  | 'others' 
  | 'all';

export type TaskStatus = 'my-tasks' | 'in-progress' | 'completed';

export interface Task {
  id: string;
  name: string;
  category: TaskCategory;
  dueDate: string;
  description: string;
  status: TaskStatus;
  createdAt: string;
  isAdminTask: boolean;
}

export type NewTask = Omit<Task, 'id' | 'createdAt' | 'isAdminTask'>;